refactor: extract CommandLineRunner
Each of the out-of-process parser classes implemented its own
CommandLineRunner, and they were all functionally the same (with some
small differences that weren't actually used). This pulls the class up
to one reused class.
