+++ lib/cc/engine/analyzers/php/parser.rb
@@ -1,5 +1,6 @@
-require 'cc/engine/analyzers/php/ast'
-require 'cc/engine/analyzers/php/nodes'
+require "cc/engine/analyzers/command_line_runner"
+require "cc/engine/analyzers/php/ast"
+require "cc/engine/analyzers/php/nodes"
 
 module CC
   module Engine
@@ -34,39 +35,6 @@ module CC
             File.expand_path(
               File.join(File.dirname(__FILE__), relative_path)
             )
-          end
-        end
-
-        class CommandLineRunner
-          attr_reader :command, :delegate
-
-          DEFAULT_TIMEOUT = 20
-
-          def initialize(command)
-            @command = command
-          end
-
-          def run(input, timeout = DEFAULT_TIMEOUT)
-            Timeout.timeout(timeout) do
-              Open3.popen3 command, "r+" do |stdin, stdout, stderr, wait_thr|
-                stdin.puts input
-                stdin.close
-
-                exit_code = wait_thr.value
-
-                output = stdout.gets
-                stdout.close
-
-                err_output = stderr.gets
-                stderr.close
-
-                if 0 == exit_code
-                  yield output
-                else
-                  raise ::CC::Engine::Analyzers::ParserError, "Python parser exited with code #{exit_code}:\n#{err_output}"
-                end
-              end
-            end
           end
         end
       end
